@@ -16,7 +16,7 @@ if (POLICY CMP0075)
 endif()
 
 # Use C++17
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 
 # Require C++17
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
@@ -40,8 +40,6 @@ set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_CURRENT_SOURCE_DIR})
 # Find packages
 find_package(DOLFINX 0.4.2.0 REQUIRED)
 find_package(Basix 0.4.2.0 REQUIRED)
-find_package(xtensor REQUIRED)
-find_package(DOLFINX_CUAS 0.4.0 REQUIRED)
 
 feature_summary(WHAT ALL)
 
@@ -55,16 +53,20 @@ target_link_libraries(dolfinx_contact PUBLIC Basix::basix)
 target_link_libraries(dolfinx_contact PUBLIC dolfinx)
 
 include(GNUInstallDirs)
-install(FILES Contact.h contact_kernels.hpp utils.h coefficients.h geometric_quantities.h SubMesh.h QuadratureRule.h RayTracing.h DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/dolfinx_contact COMPONENT Development)
+install(FILES Contact.h contact_kernels.hpp error_handling.h utils.h coefficients.h elasticity.h geometric_quantities.h meshtie_kernels.h SubMesh.h QuadratureRule.h RayTracing.h KernelData.h DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/dolfinx_contact COMPONENT Development)
 
 target_sources(dolfinx_contact PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/utils.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/coefficients.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/elasticity.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/geometric_quantities.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/meshtie_kernels.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/SubMesh.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/QuadratureRule.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/Contact.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/RayTracing.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/KernelData.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/error_handling.cpp
   )
 
 # Set target include location (for build and installed)
