@@ -5,7 +5,7 @@ project(${PROJECT_NAME})
 
 
 # Set C++ standard before finding pybind11
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
