@@ -22,88 +22,88 @@ dolfinx_contact::QuadratureRule::QuadratureRule(dolfinx::mesh::CellType ct,
 
   basix::cell::type b_ct = dolfinx::mesh::cell_type_to_basix_type(ct);
   _num_sub_entities = basix::cell::num_sub_entities(b_ct, dim);
-  const int tdim = basix::cell::topological_dimension(b_ct);
-
+  _tdim = basix::cell::topological_dimension(b_ct);
+  assert(dim <= 3);
   // If cell dimension no pushing forward
-  if (tdim == dim)
+  if (_tdim == std::size_t(dim))
   {
-    auto [q_points, q_weights]
+    std::array<std::vector<double>, 2> quadrature
         = basix::quadrature::make_quadrature(type, b_ct, degree);
+    std::vector<double>& q_weights = quadrature.back();
+    std::size_t num_points = q_weights.size();
+    std::size_t pt_shape = quadrature.front().size() / num_points;
+    cmdspan2_t qp(quadrature.front().data(), num_points, pt_shape);
 
-    std::size_t num_points = q_points.shape(0);
-    _points = xt::empty<double>(
-        {std::size_t(num_points * _num_sub_entities), (std::size_t)tdim});
-    _entity_offset = std::vector<std::int32_t>(_num_sub_entities + 1, 0);
-    std::vector<double> weights(num_points * _num_sub_entities);
+    _points = std::vector<double>(num_points * _num_sub_entities * _tdim);
+    _entity_offset = std::vector<std::size_t>(_num_sub_entities + 1, 0);
+    _weights = std::vector<double>(num_points * _num_sub_entities);
     for (std::int32_t i = 0; i < _num_sub_entities; i++)
     {
-      _entity_offset[i + 1] = (i + 1) * (std::int32_t)q_weights.size();
+      _entity_offset[i + 1] = (i + 1) * q_weights.size();
       for (std::size_t j = 0; j < num_points; ++j)
       {
-        weights[i * num_points * _num_sub_entities + j] = q_weights[j];
-        for (int k = 0; k < tdim; ++k)
-          _points[i * num_points * _num_sub_entities + j * tdim + k]
-              = q_points(j, k);
+        _weights[i * num_points * _num_sub_entities + j] = q_weights[j];
+        for (std::size_t k = 0; k < _tdim; ++k)
+          _points[i * num_points * _num_sub_entities + j * _tdim + k]
+              = qp(j, k);
       }
     }
-    _weights = weights;
   }
   else
   {
     // Create reference topology and geometry
     auto entity_topology = basix::cell::topology(b_ct)[dim];
-    const xt::xtensor<double, 2> ref_geom = basix::cell::geometry(b_ct);
 
     // Create map for each facet type to the local index
-    std::vector<xt::xarray<double>> quadrature_points;
-    std::vector<std::vector<double>> quadrature_weights;
-    quadrature_points.reserve(_num_sub_entities);
-    quadrature_weights.reserve(_num_sub_entities);
-    std::vector<std::int32_t> num_points_per_entity(_num_sub_entities);
+    std::vector<std::size_t> num_points_per_entity(_num_sub_entities);
     for (std::int32_t i = 0; i < _num_sub_entities; i++)
     {
+      // Create reference element to map facet quadrature to
       basix::cell::type et = basix::cell::sub_entity_type(b_ct, dim, i);
       basix::FiniteElement entity_element
           = basix::create_element(basix::element::family::P, et, 1,
                                   basix::element::lagrange_variant::gll_warped);
       // Create quadrature and tabulate on entity
-      auto [q_points, q_weights]
-          = basix::quadrature::make_quadrature(et, degree);
-      num_points_per_entity[i] = (std::int32_t)q_weights.size();
+      std::array<std::vector<double>, 2> quadrature
+          = basix::quadrature::make_quadrature(type, et, degree);
+      const std::vector<double>& q_weights = quadrature.back();
+      const std::vector<double>& q_points = quadrature.front();
+      const std::size_t num_points = q_weights.size();
+      const std::size_t tdim = q_points.size() / q_weights.size();
+      num_points_per_entity[i] = num_points;
 
-      auto c_tab = entity_element.tabulate(0, q_points);
-      xt::xtensor<double, 2> phi_s
-          = xt::view(c_tab, 0, xt::all(), xt::all(), 0);
-      auto cell_entity = entity_topology[i];
-      auto coords = xt::view(ref_geom, xt::keep(cell_entity), xt::all());
+      const std::array<std::size_t, 4> e_tab_shape
+          = entity_element.tabulate_shape(0, num_points);
+      std::vector<double> reference_entity_b(std::reduce(
+          e_tab_shape.cbegin(), e_tab_shape.cend(), 1, std::multiplies{}));
+
+      entity_element.tabulate(0, q_points, {num_points, tdim},
+                              reference_entity_b);
+
+      cmdspan4_t basis_full(reference_entity_b.data(), e_tab_shape);
+      auto phi = stdex::submdspan(basis_full, 0, stdex::full_extent,
+                                  stdex::full_extent, 0);
+
+      auto [sub_geomb, sub_geom_shape]
+          = basix::cell::sub_entity_geometry(b_ct, dim, i);
+      cmdspan2_t coords(sub_geomb.data(), sub_geom_shape);
 
       // Push forward quadrature point from reference entity to reference
       // entity on cell
-      const size_t num_quadrature_pts = q_weights.size();
-      xt::xtensor<double, 2> entity_qp
-          = xt::zeros<double>({num_quadrature_pts, ref_geom.shape(1)});
-      dolfinx::math::dot(phi_s, coords, entity_qp);
-
-      quadrature_points.push_back(entity_qp);
-      quadrature_weights.push_back(q_weights);
+      const std::size_t offset = _points.size();
+      _points.resize(_points.size() + num_points * coords.extent(1));
+      mdspan2_t entity_qp(_points.data() + offset, num_points,
+                          coords.extent(1));
+      assert(coords.extent(1) == _tdim);
+      dolfinx::math::dot(phi, coords, entity_qp);
+      const std::size_t weights_offset = _weights.size();
+      _weights.resize(_weights.size() + q_weights.size());
+      std::copy(q_weights.cbegin(), q_weights.cend(),
+                std::next(_weights.begin(), weights_offset));
     }
-    _entity_offset = std::vector<std::int32_t>(_num_sub_entities + 1, 0);
+    _entity_offset = std::vector<std::size_t>(_num_sub_entities + 1, 0);
     std::partial_sum(num_points_per_entity.begin(), num_points_per_entity.end(),
                      std::next(_entity_offset.begin()));
-    _points = xt::empty<double>(
-        {(std::size_t)_entity_offset.back(), (std::size_t)tdim});
-    std::vector<double> weights(_entity_offset.back());
-    for (std::int32_t i = 0; i < _num_sub_entities; i++)
-    {
-      const std::int32_t num_points = _entity_offset[i + 1] - _entity_offset[i];
-      for (std::size_t j = 0; j < (std::size_t)num_points; ++j)
-      {
-        weights[i * num_points + j] = quadrature_weights[i][j];
-        for (int k = 0; k < tdim; ++k)
-          _points(i * num_points + j, k) = quadrature_points[i](j, k);
-      }
-    }
-    _weights = weights;
   }
 }
 
@@ -121,15 +121,24 @@ int QuadratureRule::degree() const { return _degree; }
 //-----------------------------------------------------------------------------------------------
 basix::quadrature::type QuadratureRule::type() const { return _type; }
 //-----------------------------------------------------------------------------------------------
-std::int32_t QuadratureRule::num_points(int i) const
+std::size_t QuadratureRule::num_points(int i) const
 {
   assert(i < _num_sub_entities);
   return _entity_offset[i + 1] - _entity_offset[i];
 }
 //-----------------------------------------------------------------------------------------------
-xt::xtensor<double, 2> QuadratureRule::points(int i) const
+cmdspan2_t QuadratureRule::points(int i) const
 {
   assert(i < _num_sub_entities);
-  return xt::view(_points, xt::xrange(_entity_offset[i], _entity_offset[i + 1]),
-                  xt::all());
+  cmdspan2_t all_points(_points.data(), _weights.size(), _tdim);
+  return stdex::submdspan(all_points,
+                          std::pair(_entity_offset[i], _entity_offset[i + 1]),
+                          stdex::full_extent);
+}
+//-----------------------------------------------------------------------------------------------
+std::span<const double> QuadratureRule::weights(int i) const
+{
+  assert(i < _num_sub_entities);
+  return std::span(_weights.data() + _entity_offset[i],
+                   _entity_offset[i + 1] - _entity_offset[i]);
 }
