@@ -13,6 +13,22 @@
 
 namespace dolfinx_contact
 {
+namespace stdex = std::experimental;
+using mdspan2_t = stdex::mdspan<double, stdex::dextents<std::size_t, 2>>;
+using cmdspan2_t = stdex::mdspan<const double, stdex::dextents<std::size_t, 2>>;
+using mdspan3_t = stdex::mdspan<double, stdex::dextents<std::size_t, 3>>;
+using cmdspan3_t = stdex::mdspan<const double, stdex::dextents<std::size_t, 3>>;
+using mdspan4_t = stdex::mdspan<double, stdex::dextents<std::size_t, 4>>;
+using cmdspan4_t = stdex::mdspan<const double, stdex::dextents<std::size_t, 4>>;
+using s_cmdspan2_t
+    = stdex::mdspan<const double, stdex::dextents<std::size_t, 2>,
+                    stdex::layout_stride>;
+using s_cmdspan3_t
+    = stdex::mdspan<const double, stdex::dextents<std::size_t, 3>,
+                    stdex::layout_stride>;
+using s_cmdspan4_t
+    = stdex::mdspan<const double, stdex::dextents<std::size_t, 4>,
+                    stdex::layout_stride>;
 
 class QuadratureRule
 {
@@ -29,7 +45,7 @@ public:
                  = basix::quadrature::type::Default);
 
   /// Return a list of quadrature points for each entity in the cell
-  const xt::xtensor<double, 2>& points() const { return _points; }
+  const std::vector<double>& points() const { return _points; }
 
   /// Return a list of quadrature weights for each entity in the cell (using
   /// local entity index as in DOLFINx/Basix)
@@ -53,25 +69,34 @@ public:
   basix::quadrature::type type() const;
 
   /// Return the number of quadrature points per entity
-  std::int32_t num_points(int i) const;
+  std::size_t num_points(int i) const;
+
+  /// Return the topological dimension of the quadrature rule
+  std::size_t tdim() const { return _tdim; };
 
   /// Return the quadrature points for the ith entity
   /// @param[in] i The local entity index
-  xt::xtensor<double, 2> points(int i) const;
+  cmdspan2_t points(int i) const;
+
+  /// Return the quadrature weights for the ith entity
+  /// @param[in] i The local entity index
+  std::span<const double> weights(int i) const;
 
   /// Return offset for quadrature rule of the ith entity
-  const std::vector<std::int32_t>& offset() const { return _entity_offset; }
+  const std::vector<std::size_t>& offset() const { return _entity_offset; }
 
 private:
   dolfinx::mesh::CellType _cell_type;
+  std::size_t _tdim;
   int _degree;
   basix::quadrature::type _type;
   int _dim;
-  xt::xtensor<double, 2>
-      _points; // Quadrature points for each entity on the cell
+  std::vector<double>
+      _points; // Quadrature points for each entity on the cell. Shape (entity,
+               // num_points, tdim). Flattened row-major.
   std::vector<double>
       _weights; // Quadrature weights for each entity on the cell
-  std::vector<std::int32_t> _entity_offset; // The offset for each entity
+  std::vector<std::size_t> _entity_offset; // The offset for each entity
 
   int _num_sub_entities; // Number of sub entities
 };
