@@ -7,40 +7,49 @@
 
 #include "RayTracing.h"
 //------------------------------------------------------------------------------------------------
-std::tuple<int, std::int32_t, xt::xtensor<double, 1>, xt::xtensor<double, 1>>
-dolfinx_contact::raytracing(
-    const dolfinx::mesh::Mesh& mesh, const xt::xtensor<double, 1>& point,
-    const xt::xtensor<double, 1>& normal,
-    const std::vector<std::pair<std::int32_t, int>>& cells, const int max_iter,
-    const double tol)
+std::tuple<int, std::int32_t, std::vector<double>, std::vector<double>>
+dolfinx_contact::raytracing(const dolfinx::mesh::Mesh& mesh,
+                            std::span<const double> point,
+                            std::span<const double> normal,
+                            std::span<const std::int32_t> cells,
+                            const int max_iter, const double tol)
 {
   const int tdim = mesh.topology().dim();
   const int gdim = mesh.geometry().dim();
 
-  assert((std::size_t)gdim == point.shape(0));
-  assert(normal.shape(0) == (std::size_t)gdim);
-  std::tuple<int, std::int32_t, xt::xtensor<double, 1>, xt::xtensor<double, 1>>
+  assert((std::size_t)gdim == point.size());
+  assert((std::size_t)gdim == normal.size());
+  std::tuple<int, std::int32_t, std::vector<double>, std::vector<double>>
       output;
   if (tdim == 2)
   {
     if (gdim == 2)
     {
       auto [status, cell_idx, x, X] = dolfinx_contact::compute_ray<2, 2>(
-          mesh, point, normal, cells, max_iter, tol);
-      output = std::make_tuple(status, cell_idx, x, X);
+          mesh, std::span<const double, 2>(point.data(), 2),
+          std::span<const double, 2>(normal.data(), 2), cells, max_iter, tol);
+      std::vector<double> x_out(x.begin(), x.end());
+      std::vector<double> X_out(X.begin(), X.end());
+      output = std::make_tuple(status, cell_idx, x_out, X_out);
     }
     else if (gdim == 3)
     {
       auto [status, cell_idx, x, X] = dolfinx_contact::compute_ray<2, 3>(
-          mesh, point, normal, cells, max_iter, tol);
-      output = std::make_tuple(status, cell_idx, x, X);
+          mesh, std::span<const double, 3>(point.data(), 3),
+          std::span<const double, 3>(normal.data(), 3), cells, max_iter, tol);
+      std::vector<double> x_out(x.begin(), x.end());
+      std::vector<double> X_out(X.begin(), X.end());
+      output = std::make_tuple(status, cell_idx, x_out, X_out);
     }
   }
   else if (tdim == 3)
   {
     auto [status, cell_idx, x, X] = dolfinx_contact::compute_ray<3, 3>(
-        mesh, point, normal, cells, max_iter, tol);
-    output = std::make_tuple(status, cell_idx, x, X);
+        mesh, std::span<const double, 3>(point.data(), 3),
+        std::span<const double, 3>(normal.data(), 3), cells, max_iter, tol);
+    std::vector<double> x_out(x.begin(), x.end());
+    std::vector<double> X_out(X.begin(), X.end());
+    output = std::make_tuple(status, cell_idx, x_out, X_out);
   }
   return output;
 }
