@@ -13,7 +13,6 @@
 #include <dolfinx/fem/utils.h>
 #include <dolfinx/mesh/Mesh.h>
 #include <dolfinx/mesh/MeshTags.h>
-#include <xtensor/xadapt.hpp>
 
 namespace dolfinx_contact
 {
@@ -29,9 +28,9 @@ public:
   /// @param[in] mesh - the parent mesh
   /// @param[in] facets - vector of pairs (cell, facet) of exterior facets,
   /// where cell is the index of the cell local to the process and facet is
-  /// the facet index within the cell
+  /// the facet index within the cell. The data is flattened row-major.
   SubMesh(std::shared_ptr<const dolfinx::mesh::Mesh> mesh,
-          std::vector<std::pair<std::int32_t, int>>& facets);
+          std::span<const std::int32_t> facets);
 
   // Return mesh
   std::shared_ptr<const dolfinx::mesh::Mesh> mesh() const { return _mesh; }
@@ -53,14 +52,18 @@ public:
 
   // Return parent cells: parent_cells()[i] is the cell in the parent mesh for
   // ith cell in submesh
-  std::vector<std::int32_t> parent_cells() const { return _parent_cells; }
+  std::span<const std::int32_t> parent_cells() const
+  {
+    return std::span<const std::int32_t>(_parent_cells.data(),
+                                         _parent_cells.size());
+  }
 
   // Create FunctionSpace on submesh that is identical with a given
   // FunctionSpace on the parent mesh but restricted to submesh
   /// @param[in] V_parent - the function space on the the parent mesh
   /// @return the function space on the submesh
   dolfinx::fem::FunctionSpace create_functionspace(
-      std::shared_ptr<dolfinx::fem::FunctionSpace> V_parent) const;
+      std::shared_ptr<const dolfinx::fem::FunctionSpace> V_parent) const;
 
   // Copy of a function on the parent mesh/ in the parent function space
   // to submesh/ function space on submesh
@@ -69,6 +72,19 @@ public:
   /// copied
   void copy_function(dolfinx::fem::Function<PetscScalar>& u_parent,
                      dolfinx::fem::Function<PetscScalar>& u_sub);
+
+  /// @brief Adds perturbation u to mesh
+  /// @param[in] u: The function to perturb the mesh with. The function must be
+  /// based on the same finite element as the mesh coordinate element.
+  void update_geometry(dolfinx::fem::Function<PetscScalar>& u);
+
+  /// Map parent facets (parent_cell, local_facet_index) to submesh (cell,
+  /// local_facet_index) tuples
+  /// @param[in] tag: The submesh tag
+  /// @param[in] facets: The facets. Flattened row-major
+  /// @returns The submesh facets pairs. Flattened row-major
+  std::vector<std::int32_t>
+  get_submesh_tuples(int i, std::span<const std::int32_t> facets) const;
 
 private:
   // the submesh mesh
