@@ -7,7 +7,9 @@
 #pragma once
 
 #include "Contact.h"
+#include "KernelData.h"
 #include "QuadratureRule.h"
+#include "elasticity.h"
 #include "geometric_quantities.h"
 #include "utils.h"
 #include <basix/cell.h>
@@ -29,95 +31,43 @@ kernel_fn<T> generate_contact_kernel(
   assert(mesh);
 
   // Get mesh info
-  const dolfinx::mesh::Geometry& geometry = mesh->geometry();
-  const dolfinx::fem::CoordinateElement& cmap = geometry.cmap();
-  const int num_coordinate_dofs = cmap.dim();
-
-  const std::uint32_t gdim = geometry.dim();
-  const dolfinx::mesh::Topology& topology = mesh->topology();
-  const std::uint32_t tdim = topology.dim();
-  const dolfinx::mesh::CellType ct = topology.cell_type();
-  const int fdim = tdim - 1;
-
-  if ((ct == dolfinx::mesh::CellType::prism)
-      or (ct == dolfinx::mesh::CellType::pyramid))
-  {
-    throw std::invalid_argument("Unsupported cell type");
-  }
+  const std::uint32_t gdim = mesh->geometry().dim();
+  const std::uint32_t tdim = mesh->topology().dim();
 
   // Create quadrature points on reference facet
-  const std::vector<double>& q_weights = quadrature_rule.weights();
-  const xt::xtensor<double, 2>& q_points = quadrature_rule.points();
-  const std::size_t num_quadrature_pts = q_weights.size();
-
-  // Structures needed for basis function tabulation
-  // phi and grad(phi) at quadrature points
-  std::shared_ptr<const dolfinx::fem::FiniteElement> element = V->element();
-  const std::uint32_t bs = element->block_size();
-  std::uint32_t ndofs_cell = element->space_dimension() / bs;
-  if (element->value_size() / bs != 1)
-  {
-    throw std::invalid_argument(
-        "Contact kernel not supported for spaces with value size!=1");
-  }
-  if (bs != gdim)
-  {
-    throw std::invalid_argument("Contact kernel not supported for vector "
-                                "spaces of other dimension than gdim");
-  }
-  /// Pack test and trial functions
-  xt::xtensor<double, 2> phi({num_quadrature_pts, ndofs_cell});
-  xt::xtensor<double, 3> dphi({tdim, num_quadrature_pts, ndofs_cell});
-  {
-    xt::xtensor<double, 4> basis_functions(
-        {(std::size_t(tdim + 1), num_quadrature_pts, (std::size_t)ndofs_cell,
-          1)});
-    element->tabulate(basis_functions, q_points, 1);
-    phi = xt::view(basis_functions, 0, xt::all(), xt::all(), 0);
-    dphi = xt::view(basis_functions, xt::range(1, tdim + 1), xt::all(),
-                    xt::all(), 0);
-  }
-
-  // Tabulate Coordinate element (first derivative to compute Jacobian)
-  std::array<std::size_t, 4> cmap_shape
-      = cmap.tabulate_shape(1, q_weights.size());
-  xt::xtensor<double, 3> dphi_c({tdim, cmap_shape[1], cmap_shape[2]});
-  {
-    xt::xtensor<double, 4> cmap_basis(cmap_shape);
-    cmap.tabulate(1, q_points, cmap_basis);
-    dphi_c
-        = xt::view(cmap_basis, xt::range(1, tdim + 1), xt::all(), xt::all(), 0);
-  }
+  const std::vector<double>& q_points = quadrature_rule.points();
+  assert(quadrature_rule.tdim() == tdim);
+  const std::size_t num_quadrature_pts = quadrature_rule.weights().size();
 
   // Structures for coefficient data
   int num_coeffs = coeffs.size();
-  std::vector<int> offsets(num_coeffs + 4);
-  offsets[0] = 0;
-  for (int i = 1; i < num_coeffs + 1; i++)
+  std::vector<std::size_t> cstrides(num_coeffs + 3);
+  for (int i = 0; i < num_coeffs; i++)
   {
     std::shared_ptr<const dolfinx::fem::FiniteElement> coeff_element
-        = coeffs[i - 1]->function_space()->element();
-    offsets[i]
-        = offsets[i - 1]
-          + coeff_element->space_dimension() / coeff_element->block_size();
+        = coeffs[i]->function_space()->element();
+    cstrides[i]
+        = coeff_element->space_dimension() / coeff_element->block_size();
   }
 
   // FIXME: This will not work for prism meshes
-  const std::vector<std::int32_t>& qp_offsets = quadrature_rule.offset();
+  const std::vector<std::size_t>& qp_offsets = quadrature_rule.offset();
   const std::size_t num_qp_per_entity = qp_offsets[1] - qp_offsets[0];
-  offsets[num_coeffs + 1] = offsets[num_coeffs] + 1; // h
-  offsets[num_coeffs + 2]
-      = offsets[num_coeffs + 1] + gdim * num_qp_per_entity; // gap
-  offsets[num_coeffs + 3]
-      = offsets[num_coeffs + 2] + gdim * num_qp_per_entity; // normals
+  cstrides[num_coeffs] = 1;                            // h
+  cstrides[num_coeffs + 1] = gdim * num_qp_per_entity; // gap
+  cstrides[num_coeffs + 2] = gdim * num_qp_per_entity; // normals
 
-  // Pack coefficients for functions and gradients of functions (untested)
-  // FIXME: This assumption would fail for prisms
+  dolfinx_contact::KernelData kd(
+      V, std::make_shared<dolfinx_contact::QuadratureRule>(quadrature_rule),
+      cstrides);
 
-  // Tabulate basis functions and first derivatives for all input coefficients
-  xt::xtensor<double, 2> phi_coeffs({num_quadrature_pts, offsets[num_coeffs]});
-  xt::xtensor<double, 3> dphi_coeffs(
-      {tdim, num_quadrature_pts, offsets[num_coeffs]});
+  // Tabulate basis functions for all input
+  // coefficients
+  std::vector<double> coeff_basis_valuesb(num_quadrature_pts
+                                          * kd.offsets(num_coeffs));
+  const std::array<std::size_t, 2> coeff_shape
+      = {num_quadrature_pts, kd.offsets(num_coeffs)};
+  mdspan2_t coeff_basis_valuesm(coeff_basis_valuesb.data(), coeff_shape);
 
   // Create finite elements for coefficient functions and tabulate shape
   // functions
@@ -125,36 +75,26 @@ kernel_fn<T> generate_contact_kernel(
   {
     std::shared_ptr<const dolfinx::fem::FiniteElement> coeff_element
         = coeffs[i]->function_space()->element();
-    xt::xtensor<double, 4> coeff_basis(
-        {tdim + 1, num_quadrature_pts,
-         coeff_element->space_dimension() / coeff_element->block_size(), 1});
-    if (coeff_element->value_size() / coeff_element->block_size() != 1)
+    assert(coeff_element);
+    const std::array<std::size_t, 4> coeff_i_shape
+        = coeff_element->basix_element().tabulate_shape(0, num_quadrature_pts);
+    if (coeff_i_shape.back() != 1)
     {
       throw std::invalid_argument(
           "Kernel does not support coefficients with value size!=1");
     }
-    coeff_element->tabulate(coeff_basis, q_points, 1);
-    auto phi_i = xt::view(phi_coeffs, xt::all(),
-                          xt::range(offsets[i], offsets[i + 1]));
-    phi_i = xt::view(coeff_basis, 0, xt::all(), xt::all(), 0);
-    auto dphi_i = xt::view(dphi_coeffs, xt::all(), xt::all(),
-                           xt::range(offsets[i], offsets[i + 1]));
-    dphi_i = xt::view(coeff_basis, xt::range(1, tdim + 1), xt::all(), xt::all(),
-                      0);
+    std::vector<double> coeff_basisb(std::reduce(
+        coeff_i_shape.cbegin(), coeff_i_shape.cend(), 1, std::multiplies{}));
+    coeff_element->tabulate(coeff_basisb, q_points, {num_quadrature_pts, tdim},
+                            0);
+    cmdspan4_t coeff_basis(coeff_basisb.data(), coeff_i_shape);
+    auto basis = stdex::submdspan(coeff_basis_valuesm, stdex::full_extent,
+                                  std::pair{kd.offsets(i), kd.offsets(i + 1)});
+    assert(kd.offsets(i + 1) - kd.offsets(i) == coeff_i_shape[2]);
+    for (std::size_t k = 0; k < basis.extent(0); ++k)
+      for (std::size_t l = 0; l < basis.extent(1); ++l)
+        basis(k, l) = coeff_basis(0, k, l, 0);
   }
-  // As reference facet and reference cell are affine, we do not need to compute
-  // this per quadrature point
-  auto ref_jacobians = basix::cell::facet_jacobians(
-      dolfinx::mesh::cell_type_to_basix_type(ct));
-
-  // Get facet normals on reference cell
-  auto facet_normals = basix::cell::facet_outward_normals(
-      dolfinx::mesh::cell_type_to_basix_type(ct));
-
-  auto update_jacobian
-      = dolfinx_contact::get_update_jacobian_dependencies(cmap);
-  auto update_normal = dolfinx_contact::get_update_normal(cmap);
-  const bool affine = cmap.is_affine();
 
   /// @brief Kernel for contact with rigid surface (RHS).
   ////
@@ -172,48 +112,47 @@ kernel_fn<T> generate_contact_kernel(
   /// facet. Used to access the correct quadrature rule.
   /// @param[in] num_links Unused integer. In two sided contact this indicates
   /// how many cells are connected with the cell.
+  /// @param[in] q_indices Unused indices. In two sided contact this yields what
+  /// quadrature points to add contributions from
   dolfinx_contact::kernel_fn<T> nitsche_rigid_rhs
-      = [=](std::vector<std::vector<T>>& b, const T* c, const T* w,
+      = [kd, gdim, tdim, coeff_basis_valuesb, coeff_shape, constant_normal](
+            std::vector<std::vector<T>>& b, std::span<const T> c, const T* w,
             const double* coordinate_dofs, const int facet_index,
-            [[maybe_unused]] const std::size_t num_links)
+            [[maybe_unused]] const std::size_t num_links,
+            [[maybe_unused]] std::span<const std::int32_t> q_indices)
   {
-    // assumption that the vector function space has block size tdim
-    std::array<std::int32_t, 2> q_offset
-        = {qp_offsets[facet_index], qp_offsets[facet_index + 1]};
+    // Retrieve some data from kd
+    const std::array<std::int32_t, 2> q_offset
+        = {kd.qp_offsets(facet_index), kd.qp_offsets(facet_index + 1)};
+    const std::size_t bs = kd.bs();
+    const std::uint32_t ndofs_cell = kd.ndofs_cell();
 
-    // Reshape coordinate dofs to two dimensional array
-    // FIXME: These array should be views (when compute_jacobian doesn't use
-    // xtensor)
-    std::array<std::size_t, 2> shape = {num_coordinate_dofs, 3};
-    const xt::xtensor<double, 2>& coord = xt::adapt(
-        coordinate_dofs, num_coordinate_dofs * 3, xt::no_ownership(), shape);
-    auto c_view = xt::view(coord, xt::all(), xt::range(0, gdim));
-
-    // Extract the first derivative of the coordinate element (cell) of degrees
-    // of freedom on the facet
-    const xt::xtensor<double, 3> dphi_fc = xt::view(
-        dphi_c, xt::all(), xt::xrange(q_offset[0], q_offset[1]), xt::all());
+    // Reshape coordinate dofs to two-dimensional array
+    cmdspan2_t coord(coordinate_dofs, kd.num_coordinate_dofs(), 3);
 
     // Compute Jacobian and determinant at first quadrature point
-    xt::xtensor<double, 2> J = xt::zeros<double>({gdim, tdim});
-    xt::xtensor<double, 2> K = xt::zeros<double>({tdim, gdim});
-    xt::xtensor<double, 2> J_f
-        = xt::view(ref_jacobians, facet_index, xt::all(), xt::all());
-    xt::xtensor<double, 2> J_tot
-        = xt::zeros<double>({J.shape(0), J_f.shape(1)});
+    std::array<double, 9> Jb;
+    mdspan2_t J(Jb.data(), gdim, tdim);
+    std::array<double, 9> Kb;
+    mdspan2_t K(Kb.data(), tdim, gdim);
+    std::array<double, 6> J_totb;
+    mdspan2_t J_tot(J_totb.data(), gdim, tdim - 1);
+    double detJ = 0;
+    std::array<double, 18> detJ_scratch;
 
-    double detJ;
     // Normal vector on physical facet at a single quadrature point
-    xt::xtensor<double, 1> n_phys = xt::zeros<double>({gdim});
-    // Pre-compute jacobians and normals for affine meshes
-    if (affine)
-    {
-      detJ = std::fabs(dolfinx_contact::compute_facet_jacobians(
-          0, J, K, J_tot, J_f, dphi_fc, coord));
-      dolfinx_contact::physical_facet_normal(
-          n_phys, K, xt::row(facet_normals, facet_index));
-    }
+    std::array<double, 3> n_phys;
 
+    // Pre-compute jacobians and normals for affine meshes
+    if (kd.affine())
+    {
+      detJ = kd.compute_first_facet_jacobian(facet_index, J, K, J_tot,
+                                             detJ_scratch, coord);
+      dolfinx_contact::physical_facet_normal(
+          std::span(n_phys.data(), gdim), K,
+          stdex::submdspan(kd.facet_normals(), facet_index,
+                           stdex::full_extent));
+    }
     // Retrieve normal of rigid surface if constant
     std::array<double, 3> n_surf = {0, 0, 0};
     double n_dot = 0;
@@ -226,44 +165,50 @@ kernel_fn<T> generate_contact_kernel(
         // (-n_y)* (Pi(x) - x), where n_y is the outward unit normal
         // in y = Pi(x)
         n_surf[i] = -w[i + 2];
-        n_dot += n_phys(i) * n_surf[i];
+        n_dot += n_phys[i] * n_surf[i];
       }
     }
-    int c_offset = (bs - 1) * offsets[1];
+    int c_offset = (bs - 1) * kd.offsets(1);
     // This is gamma/h
-    double gamma = w[0] / c[c_offset + offsets[3]];
-    double gamma_inv = c[c_offset + offsets[3]] / w[0];
+    double gamma = w[0] / c[c_offset + kd.offsets(3)];
+    double gamma_inv = c[c_offset + kd.offsets(3)] / w[0];
     double theta = w[1];
-    xtl::span<const double> _weights(q_weights);
-    auto weights = _weights.subspan(q_offset[0], q_offset[1] - q_offset[0]);
-
-    // Temporary variable for grad(phi) on physical cell
-    xt::xtensor<double, 2> dphi_phys({bs, ndofs_cell});
+    std::span<const double> _weights(kd.q_weights());
+    const int num_points = q_offset.back() - q_offset.front();
+    auto weights = _weights.subspan(q_offset.front(), num_points);
 
     // Temporary work arrays
-    xt::xtensor<double, 2> tr({std::uint32_t(offsets[1] - offsets[0]), gdim});
-    xt::xtensor<double, 2> epsn({std::uint32_t(offsets[1] - offsets[0]), gdim});
+    std::vector<double> epsnb((kd.offsets(1) - kd.offsets(0)) * gdim);
+    mdspan2_t epsn(epsnb.data(), kd.offsets(1) - kd.offsets(0), gdim);
+    std::vector<double> trb((kd.offsets(1) - kd.offsets(0)) * gdim);
+    mdspan2_t tr(trb.data(), kd.offsets(1) - kd.offsets(0), gdim);
+
+    // Extract basis for coefficients
+    cmdspan2_t phi_coeffs(coeff_basis_valuesb.data(), coeff_shape);
+
+    // Extract reference to the tabulated basis function
+    cmdspan2_t phi = kd.phi();
+    cmdspan3_t dphi = kd.dphi();
 
     // Loop over quadrature points
-    const int num_points = q_offset[1] - q_offset[0];
     for (std::size_t q = 0; q < num_points; q++)
     {
-      const std::size_t q_pos = q_offset[0] + q;
+      const std::size_t q_pos = q_offset.front() + q;
 
       // Update Jacobian and physical normal
-      detJ = std::fabs(
-          update_jacobian(q, detJ, J, K, J_tot, J_f, dphi_fc, coord));
-      update_normal(n_phys, K, facet_normals, facet_index);
+      detJ = std::fabs(kd.update_jacobian(q, facet_index, detJ, J, K, J_tot,
+                                          detJ_scratch, coord));
+      kd.update_normal(std::span(n_phys.data(), gdim), K, facet_index);
 
       double mu = 0;
-      for (int j = offsets[1]; j < offsets[2]; j++)
+      for (int j = kd.offsets(1); j < kd.offsets(2); j++)
         mu += c[j + c_offset] * phi_coeffs(q_pos, j);
       double lmbda = 0;
-      for (int j = offsets[2]; j < offsets[3]; j++)
+      for (int j = kd.offsets(2); j < kd.offsets(3); j++)
         lmbda += c[j + c_offset] * phi_coeffs(q_pos, j);
 
       // if normal not constant, get surface normal at current quadrature point
-      int normal_offset = c_offset + offsets[5];
+      int normal_offset = c_offset + kd.offsets(5);
       if (!constant_normal)
       {
         n_dot = 0;
@@ -273,46 +218,30 @@ kernel_fn<T> generate_contact_kernel(
           // (-n_y)* (Pi(x) - x), where n_y is the outward unit normal
           // in y = Pi(x)
           n_surf[i] = -c[normal_offset + q * gdim + i];
-          n_dot += n_phys(i) * n_surf[i];
+          n_dot += n_phys[i] * n_surf[i];
         }
       }
-      int gap_offset = c_offset + offsets[4];
+      int gap_offset = c_offset + kd.offsets(4);
       double gap = 0;
       for (int i = 0; i < gdim; i++)
-      {
         gap += c[gap_offset + q * gdim + i] * n_surf[i];
-      }
 
-      // precompute tr(eps(phi_j e_l)), eps(phi^j e_l)n*n2
-      std::fill(tr.begin(), tr.end(), 0);
-      std::fill(epsn.begin(), epsn.end(), 0);
-      for (int j = 0; j < offsets[1] - offsets[0]; j++)
-      {
-        for (int l = 0; l < bs; l++)
-        {
-          for (int k = 0; k < tdim; k++)
-          {
-            tr(j, l) += K(k, l) * dphi(k, q_pos, j);
-            for (int s = 0; s < gdim; s++)
-            {
-              epsn(j, l) += K(k, s) * dphi(k, q_pos, j)
-                            * (n_phys(s) * n_surf[l] + n_phys(l) * n_surf[s]);
-            }
-          }
-        }
-      }
+      compute_normal_strain_basis(epsn, tr, K, dphi, n_surf,
+                                  std::span(n_phys.data(), gdim), q_pos);
+
       // compute tr(eps(u)), epsn at q
       double tr_u = 0;
       double epsn_u = 0;
       double u_dot_nsurf = 0;
-      for (int i = 0; i < offsets[1] - offsets[0]; i++)
+      for (int i = 0; i < kd.offsets(1) - kd.offsets(0); i++)
       {
-        const std::int32_t block_index = (i + offsets[0]) * bs;
+        const std::int32_t block_index = (i + kd.offsets(0)) * bs;
         for (int j = 0; j < bs; j++)
         {
-          tr_u += c[block_index + j] * tr(i, j);
-          epsn_u += c[block_index + j] * epsn(i, j);
-          u_dot_nsurf += c[block_index + j] * n_surf[j] * phi(q_pos, i);
+          const auto c_val = c[block_index + j];
+          tr_u += c_val * tr(i, j);
+          epsn_u += c_val * epsn(i, j);
+          u_dot_nsurf += c_val * n_surf[j] * phi(q_pos, i);
         }
       }
 
@@ -353,49 +282,50 @@ kernel_fn<T> generate_contact_kernel(
   /// facet. Used to access the correct quadrature rule.
   /// @param[in] num_links Unused integer. In two sided contact this indicates
   /// how many cells are connected with the cell.
-  kernel_fn<T> nitsche_rigid_jacobian =
-      [dphi_c, phi, dphi, phi_coeffs, dphi_coeffs, offsets, num_coeffs, gdim,
-       tdim, fdim, q_weights, num_coordinate_dofs, ref_jacobians, bs,
-       facet_normals, constant_normal, affine, update_jacobian, update_normal,
-       ndofs_cell, qp_offsets](std::vector<std::vector<double>>& A, const T* c,
-                               const T* w, const double* coordinate_dofs,
-                               const int facet_index,
-                               [[maybe_unused]] const std::size_t num_links)
+  /// @param[in] q_indices Unused indices. In two sided contact this yields what
+  /// quadrature points to add contributions from
+  kernel_fn<T> nitsche_rigid_jacobian
+      = [kd, gdim, tdim, coeff_basis_valuesb, coeff_shape, constant_normal](
+            std::vector<std::vector<double>>& A, std::span<const T> c,
+            const T* w, const double* coordinate_dofs, const int facet_index,
+            [[maybe_unused]] const std::size_t num_links,
+            [[maybe_unused]] std::span<const std::int32_t> q_indices)
   {
-    std::array<std::int32_t, 2> q_offset
-        = {qp_offsets[facet_index], qp_offsets[facet_index + 1]};
+    // Retrieve some data from kd
+    const std::array<std::int32_t, 2> q_offset
+        = {kd.qp_offsets(facet_index), kd.qp_offsets(facet_index + 1)};
+    const std::size_t bs = kd.bs();
+    const std::uint32_t ndofs_cell = kd.ndofs_cell();
 
-    // Reshape coordinate dofs to two dimensional array
-    // FIXME: These array should be views (when compute_jacobian doesn't use
-    // xtensor)
-    std::array<std::size_t, 2> shape = {num_coordinate_dofs, 3};
-    xt::xtensor<double, 2> coord = xt::adapt(
-        coordinate_dofs, num_coordinate_dofs * 3, xt::no_ownership(), shape);
+    // Reshape coordinate dofs to two-dimensional array
+    cmdspan2_t coord(coordinate_dofs, kd.num_coordinate_dofs(), 3);
 
-    // Extract the first derivative of the coordinate element (cell) of degrees
-    // of freedom on the facet
-    const xt::xtensor<double, 3> dphi_fc = xt::view(
-        dphi_c, xt::all(), xt::xrange(q_offset[0], q_offset[1]), xt::all());
-    xt::xtensor<double, 2> J = xt::zeros<double>({gdim, tdim});
-    xt::xtensor<double, 2> K = xt::zeros<double>({tdim, gdim});
-    xt::xtensor<double, 1> n_phys = xt::zeros<double>({gdim});
-    xt::xtensor<double, 2> J_f
-        = xt::view(ref_jacobians, facet_index, xt::all(), xt::all());
-    xt::xtensor<double, 2> J_tot
-        = xt::zeros<double>({J.shape(0), J_f.shape(1)});
-    double detJ;
-    if (affine)
+    // Compute Jacobian and determinant at first quadrature point
+    std::array<double, 9> Jb;
+    mdspan2_t J(Jb.data(), gdim, tdim);
+    std::array<double, 9> Kb;
+    mdspan2_t K(Kb.data(), tdim, gdim);
+    std::array<double, 6> J_totb;
+    mdspan2_t J_tot(J_totb.data(), gdim, tdim - 1);
+    double detJ = 0;
+    std::array<double, 18> detJ_scratch;
+
+    // Normal vector on physical facet at a single quadrature point
+    std::array<double, 3> n_phys;
+
+    // Pre-compute jacobians and normals for affine meshes
+    if (kd.affine())
     {
-      detJ = std::fabs(dolfinx_contact::compute_facet_jacobians(
-          0, J, K, J_tot, J_f, dphi_fc, coord));
+      detJ = kd.compute_first_facet_jacobian(facet_index, J, K, J_tot,
+                                             detJ_scratch, coord);
       dolfinx_contact::physical_facet_normal(
-          n_phys, K, xt::row(facet_normals, facet_index));
+          std::span(n_phys.data(), gdim), K,
+          stdex::submdspan(kd.facet_normals(), facet_index,
+                           stdex::full_extent));
     }
 
     // Retrieve normal of rigid surface if constant
     std::array<double, 3> n_surf = {0, 0, 0};
-    // FIXME: Code duplication from previous kernel, and should be made into a
-    // lambda function
     double n_dot = 0;
     if (constant_normal)
     {
@@ -406,36 +336,51 @@ kernel_fn<T> generate_contact_kernel(
         // (-n_y)* (Pi(x) - x), where n_y is the outward unit normal
         // in y = Pi(x)
         n_surf[i] = -w[i + 2];
-        n_dot += n_phys(i) * n_surf[i];
+        n_dot += n_phys[i] * n_surf[i];
       }
     }
-    int c_offset = (bs - 1) * offsets[1];
-    double gamma
-        = w[0]
-          / c[c_offset + offsets[3]]; // This is gamma/hdouble gamma = w[0];
-    double gamma_inv = c[c_offset + offsets[3]] / w[0];
+
+    int c_offset = (bs - 1) * kd.offsets(1);
+    // This is gamma/h
+    double gamma = w[0] / c[c_offset + kd.offsets(3)];
+    double gamma_inv = c[c_offset + kd.offsets(3)] / w[0];
     double theta = w[1];
+    std::span<const double> _weights(kd.q_weights());
+    const int num_points = q_offset.back() - q_offset.front();
+    auto weights = _weights.subspan(q_offset.front(), num_points);
 
-    xtl::span<const double> _weights(q_weights);
-    auto weights = _weights.subspan(q_offset[0], q_offset[1] - q_offset[0]);
+    // Temporary work arrays
+    std::vector<double> epsnb((kd.offsets(1) - kd.offsets(0)) * gdim);
+    mdspan2_t epsn(epsnb.data(), kd.offsets(1) - kd.offsets(0), gdim);
+    std::vector<double> trb((kd.offsets(1) - kd.offsets(0)) * gdim);
+    mdspan2_t tr(trb.data(), kd.offsets(1) - kd.offsets(0), gdim);
 
-    // Get number of dofs per cell
-    // Temporary variable for grad(phi) on physical cell
-    xt::xtensor<double, 2> dphi_phys({bs, ndofs_cell});
-    xt::xtensor<double, 2> tr = xt::zeros<double>({ndofs_cell, gdim});
-    xt::xtensor<double, 2> epsn = xt::zeros<double>({ndofs_cell, gdim});
-    const std::uint32_t num_points = q_offset[1] - q_offset[0];
+    // Extract basis for coefficients
+    cmdspan2_t phi_coeffs(coeff_basis_valuesb.data(), coeff_shape);
+
+    // Extract reference to the tabulated basis function
+    cmdspan2_t phi = kd.phi();
+    cmdspan3_t dphi = kd.dphi();
+
+    // Loop over quadrature points
     for (std::size_t q = 0; q < num_points; q++)
     {
-      const std::size_t q_pos = q_offset[0] + q;
+      const std::size_t q_pos = q_offset.front() + q;
 
       // Update Jacobian and physical normal
-      detJ = std::fabs(
-          update_jacobian(q, detJ, J, K, J_tot, J_f, dphi_fc, coord));
-      update_normal(n_phys, K, facet_normals, facet_index);
+      detJ = std::fabs(kd.update_jacobian(q, facet_index, detJ, J, K, J_tot,
+                                          detJ_scratch, coord));
+      kd.update_normal(std::span(n_phys.data(), gdim), K, facet_index);
+
+      double mu = 0;
+      for (int j = kd.offsets(1); j < kd.offsets(2); j++)
+        mu += c[j + c_offset] * phi_coeffs(q_pos, j);
+      double lmbda = 0;
+      for (int j = kd.offsets(2); j < kd.offsets(3); j++)
+        lmbda += c[j + c_offset] * phi_coeffs(q_pos, j);
 
       // if normal not constant, get surface normal at current quadrature point
-      int normal_offset = c_offset + offsets[5];
+      int normal_offset = c_offset + kd.offsets(5);
       if (!constant_normal)
       {
         n_dot = 0;
@@ -445,47 +390,24 @@ kernel_fn<T> generate_contact_kernel(
           // (-n_y)* (Pi(x) - x), where n_y is the outward unit normal
           // in y = Pi(x)
           n_surf[i] = -c[normal_offset + q * gdim + i];
-          n_dot += n_phys(i) * n_surf[i];
+          n_dot += n_phys[i] * n_surf[i];
         }
       }
-      int gap_offset = c_offset + offsets[4];
+      int gap_offset = c_offset + kd.offsets(4);
       double gap = 0;
       for (int i = 0; i < gdim; i++)
         gap += c[gap_offset + q * gdim + i] * n_surf[i];
 
-      // precompute tr(eps(phi_j e_l)), eps(phi^j e_l)n*n2
-      std::fill(tr.begin(), tr.end(), 0);
-      std::fill(epsn.begin(), epsn.end(), 0);
-      for (int j = 0; j < ndofs_cell; j++)
-      {
-        for (int l = 0; l < bs; l++)
-        {
-          for (int k = 0; k < tdim; k++)
-          {
-            tr(j, l) += K(k, l) * dphi(k, q_pos, j);
-            for (int s = 0; s < gdim; s++)
-            {
-              epsn(j, l) += K(k, s) * dphi(k, q_pos, j)
-                            * (n_phys(s) * n_surf[l] + n_phys(l) * n_surf[s]);
-            }
-          }
-        }
-      }
-      double mu = 0;
-      int c_offset = (bs - 1) * offsets[1];
-      for (int j = offsets[1]; j < offsets[2]; j++)
-        mu += c[j + c_offset] * phi_coeffs(facet_index, q_pos, j);
-      double lmbda = 0;
-      for (int j = offsets[2]; j < offsets[3]; j++)
-        lmbda += c[j + c_offset] * phi_coeffs(facet_index, q_pos, j);
+      compute_normal_strain_basis(epsn, tr, K, dphi, n_surf,
+                                  std::span(n_phys.data(), gdim), q_pos);
 
       // compute tr(eps(u)), epsn at q
       double tr_u = 0;
       double epsn_u = 0;
       double u_dot_nsurf = 0;
-      for (int i = 0; i < offsets[1] - offsets[0]; i++)
+      for (int i = 0; i < kd.offsets(1) - kd.offsets(0); i++)
       {
-        const std::int32_t block_index = (i + offsets[0]) * bs;
+        const std::int32_t block_index = (i + kd.offsets(0)) * bs;
         for (int j = 0; j < bs; j++)
         {
           const auto c_val = c[block_index + j];
@@ -495,7 +417,7 @@ kernel_fn<T> generate_contact_kernel(
         }
       }
 
-      double sign_u = lmbda * tr_u * n_dot + mu * epsn_u;
+      double sign_u = (lmbda * n_dot * tr_u + mu * epsn_u);
       double Pn_u
           = dolfinx_contact::dR_minus(sign_u + gamma * (gap - u_dot_nsurf));
       const double w0 = weights[q] * detJ;
