@@ -11,9 +11,10 @@ from dolfinx_contact.cpp import (Kernel, QuadratureRule,
                                  compute_active_entities, pack_circumradius,
                                  update_geometry)
 
-from .helpers import epsilon, lame_parameters, sigma_func
+from .helpers import epsilon, lame_parameters, sigma_func, compare_matrices
 from .newton_solver import ConvergenceCriterion, NewtonSolver
+
 
 __all__ = ["NewtonSolver", "ConvergenceCriterion", "lame_parameters", "epsilon",
            "sigma_func", "Kernel", "pack_circumradius", "update_geometry",
-           "QuadratureRule", "compute_active_entities"]
+           "QuadratureRule", "compute_active_entities", "compare_matrices"]
