@@ -2,15 +2,32 @@
 #
 # SPDX-License-Identifier:    MIT
 
-from .christmas_tree import create_christmas_tree_mesh
+from .utils import convert_mesh
+from .christmas_tree import create_christmas_tree_mesh, create_christmas_tree_mesh_3D
 from .contact_meshes import (create_box_mesh_2D,
                              create_box_mesh_3D, create_circle_circle_mesh,
                              create_circle_plane_mesh,
                              create_sphere_plane_mesh, create_sphere_sphere_mesh,
                              create_cylinder_cylinder_mesh)
 from .onesided_meshes import create_disk_mesh, create_sphere_mesh
-from .utils import convert_mesh
+from .split_box import (create_split_box_2D, create_split_box_3D, create_unsplit_box_2d, create_unsplit_box_3d,
+                        vertical_line, horizontal_sine)
 
-__all__ = ["create_christmas_tree_mesh", "create_circle_plane_mesh", "create_circle_circle_mesh",
-           "create_box_mesh_2D", "create_box_mesh_3D", "create_sphere_plane_mesh", "convert_mesh",
-           "create_disk_mesh", "create_sphere_mesh", "create_sphere_sphere_mesh", "create_cylinder_cylinder_mesh"]
+__all__ = ["create_christmas_tree_mesh",
+           "create_christmas_tree_mesh_3D",
+           "create_circle_plane_mesh",
+           "create_circle_circle_mesh",
+           "create_box_mesh_2D",
+           "create_box_mesh_3D",
+           "create_sphere_plane_mesh",
+           "convert_mesh",
+           "create_disk_mesh",
+           "create_sphere_mesh",
+           "create_sphere_sphere_mesh",
+           "create_cylinder_cylinder_mesh",
+           "create_split_box_2D",
+           "create_split_box_3D",
+           "create_unsplit_box_2d",
+           "create_unsplit_box_3d",
+           "vertical_line",
+           "horizontal_sine"]
